fix: remove signed comparison warning in bluetooth prewrite
Agent-Logs-Url: https://github.com/3218158258h/gateway/sessions/3ff75ca2-3b82-458a-b0af-ca2f63ddb1e0

+++ src/app_bluetooth.c
@@ -403,7 +403,7 @@ int app_bluetooth_preWrite(Device *device, void *ptr, int *len)
     }
 
     memcpy(&temp, ptr, 1);
-    if (temp != ctx->protocol.connection_type)
+    if (temp != (int)ctx->protocol.connection_type)
     {
         *len = 0;
         return 0;
